some package renaming
