+++ flashlib.robot.systems/src/main/java/com/flash3388/flashlib/robot/systems2/MotorSystem.java
@@ -1,9 +1,8 @@
-package com.flash3388.flashlib.robot.nsys;
+package com.flash3388.flashlib.robot.systems2;
 
 import com.flash3388.flashlib.io.devices.SpeedController;
 import com.flash3388.flashlib.io.devices.SpeedControllerGroup;
-import com.flash3388.flashlib.robot.nact.Rotate;
-import com.flash3388.flashlib.robot.nint.Rotatable;
+import com.flash3388.flashlib.robot.systems2.actions.Rotate;
 import com.flash3388.flashlib.scheduling.Subsystem;
 import com.flash3388.flashlib.scheduling.actions.Action;
