+++ flashlib.robot.systems/src/main/java/com/flash3388/flashlib/robot/systems2/OmniDriveSystem.java
@@ -1,8 +1,7 @@
-package com.flash3388.flashlib.robot.nsys;
+package com.flash3388.flashlib.robot.systems2;
 
 import com.flash3388.flashlib.io.devices.SpeedController;
-import com.flash3388.flashlib.robot.nact.DriveOmni;
-import com.flash3388.flashlib.robot.nint.OmniDrive;
+import com.flash3388.flashlib.robot.systems2.actions.DriveOmni;
 import com.flash3388.flashlib.robot.systems.drive.OmniDriveSpeed;
 import com.flash3388.flashlib.robot.systems.drive.algorithms.DriveAlgorithms;
 import com.flash3388.flashlib.scheduling.Subsystem;
