+++ flashlib.robot.systems/src/main/java/com/flash3388/flashlib/robot/systems2/Rotatable.java
@@ -1,4 +1,4 @@
-package com.flash3388.flashlib.robot.nint;
+package com.flash3388.flashlib.robot.systems2;
 
 import com.beans.util.function.Suppliers;
 import com.flash3388.flashlib.control.Direction;
