+++ flashlib.robot.systems/src/main/java/com/flash3388/flashlib/robot/systems2/actions/Rotate.java
@@ -1,4 +1,4 @@
-package com.flash3388.flashlib.robot.nact;
+package com.flash3388.flashlib.robot.systems2.actions;
 
 import com.flash3388.flashlib.robot.motion.Rotatable;
 import com.flash3388.flashlib.scheduling.actions.ActionBase;
